refactor: audit 2.2 — extract LibraryHeroSection / GenreRow / PosterCard
Decomposes MediaLibraryScreen (§13 #4): 1005 → 726 lines (−28%).

- LibraryHeroSection owns the iOS browse-view hero block (backdrop,
  rating/metadata line, title, overview, Play / More info buttons) plus
  all hero-specific adaptive sizing.
- LibraryGenreRow is a thin ContentRow wrapper with a caller-owned
  "See all" callback, keeping the row viewmodel-agnostic.
- LibraryPosterCard consolidates the NavigationLink → MediaDetailScreen
  card used by the genre row and by both filtered grids.

Behavior-preserving: iOS + tvOS builds clean, all 36 tests pass in ~4.9 s.
